+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Toolium Changelog
 =================
 
+
+v1.4.2
+------
+
+*Release date: In development*
+
+- Add movement in X axis in *scroll_element_into_view* method
+- Fix bugs in  the Dynamic Environment feature:
+     - chars no utf-8 are accepted
+     - no replace behave prefixes into a step
+
 v1.4.1
 ------